[ci_stats] Add Buildkite CI support
Integrates Buildkite CI into the `ci_stats` tool, allowing it to report on
test failures from Buildkite pipelines alongside existing GitHub Actions support.

This is a first pass, at the moment it simply looks for a badge in the
README that points to a buildkite build, and then checks that. More to
come soon.

Signed-off-by: Phil Dibowitz <phil@ipom.com>