Moving towards a more standard installation system (#62)
Completely rewrites how downstream repos are setup. In short, we
bootstrap a Gemfile which points to this git repo, and install that.
From there, we use the stuff in the installed gem.

To enable this nicely, this adds oss_stats_script that will run script
from the gem's scripts directory so that we can continue to leverage
the same linters and stuff

For rubocop, I moved `inherit_from` to `gem_inherit`, and for mdl,
I just did some ruby magic to find the installed gem.

* mdl: fix config accidentally disabling all rules
* rubocop: move TargetRubyVersion to 3.2 to match rest of our config
* gemdeps: base64 and faraday-retry are direct deps, reflect that
* README: reflect new process
* initialize_repo: move to bin, write as above
* various lint things from rubocop upgrade
* add shellcheck

Signed-off-by: Phil Dibowitz <phil@ipom.com>